types: standardize on uint64_t
The problem is that on some platforms __u64 == uint64_t (x86_64), and on
others it's doesn't (ppc64).  Which means we don't know whether to define
different versions of overloaded functions for both types or just one.

So, standardize on uint64_t.  This plays nicer with STL, which defines
hash<uint64_t> on 64 bit arches but not 32 bit.  Which means we can't
standarzie on __u64 or else hash<__u64> won't work.  Bah!

Conflicts:

	src/common/Throttle.h
	src/config.h
	src/include/librados.h
	src/include/librados.hpp
	src/librados.cc
	src/messages/MPoolOp.h
	src/mon/OSDMonitor.cc
	src/os/FileJournal.cc
	src/osd/osd_types.h
	src/osdc/Objecter.h