filestore: op_start when op is _queued_, so that q is drained on commit
We need the store in a consistent state on commit, which means flushing
transactions such that we have all ops <= a given seq applied.  That is
handled by the commit_start()/commit_started() pair, but will only include
ops in the FileStore queue if we op_start when it is initially queued.
Which is exactly what we want, because the queue can reorder things, so
stopping just currently-being-applied updates will only keep transactions
atomic but not ordered.

Yup!