rgw: Support loglevel settings
The gateway now supports changing the verbosity of the logging, this can be achieved by changing the
RGW_LOG_LEVEL envirioment parameter.

For Apache running systems, this can be done with mod_env: SetEnv RGW_LOG_LEVEL 20

This will change the loglevel to the highest level and produce a lot of usefull output for debugging purposes.

For now, all the messages will go to webservers logfile (For Apache: error_log)