osd: loosen caller_ops asserts
The problem is that merge_log adds new items to the log before it unindexes
divergent items, and that behavior is needed by the current implementation
of merge_old_entry().  Since the divergent items may be the same requests
(and frequently are) these asserts needs to be loosened up.

Now, the most recent addition "wins," and we only remove the entry in
unindex() if it points to us.

Signed-off-by: Sage Weil <sage@newdream.net>