osd: fix merge_log cut point
Look at the eversion.version field (not the whole eversion) when deciding
what is divergent.  That way if we have

our log: 100'10 (0'0) m 10000004d3a.00000000/head by client4225.1:18529
new log: 122'10 (0'0) m 10000004d3a.00000000/head by client4225.1:18529

The 100'10 is divergent and the 122'10 wins and we don't get a dup
reqid in the log.

Signed-off-by: Sage Weil <sage@newdream.net>