mds: use correct helper when pinning past snaprealm parent
The heler also updates the SnapRealm::open_past_parents, which is needed
for the have_past_parents_open() check.

That is used when, among other things, we import caps; not updating it
prevented the cap import from sending the client cap message, which makes
the mds<->client cap relationship get out of sync.

Signed-off-by: Sage Weil <sage@newdream.net>