objecter: refactor interface with object_locator_t
This paves the way for a locator that lets the user specify an arbitrary
string to hash for placement (instead of the object name).  It also
captures everything that would affect placement (pool, preferred, etc.)
in a single type.

Just the client side; no server side or protocol changes yet.

Signed-off-by: Sage Weil <sage@newdream.net>