OSD::_process_pg_info:search_for_missing sometimes
OSD::_process_pg_info: If we're the primary for this active PG, and we
have missing objects, call search_for_missing. This should ensure that
we know where to find our missing objects.

The reason why this wasn't there before is that previously, we kept the
PG from going active until all the missing objects were found.

Signed-off-by: Colin McCabe <colinm@hq.newdream.net>