osd: drop not-quite-copy constructor for object_info_t
Making a copy-like constructor that doesn't actaully copy is confusing
and error prone.  In this case, we initialized a clone's object_info with
the head's snapid, causing problems with what info was encoded and crashing
later in the snap_trimmer.  Here the one caller already called
copy_user_bits(); let's move the lost copy there.

This backs out one of the changes in 0cc8d34e.

Signed-off-by: Sage Weil <sage@newdream.net>