mds: fix replay/resent vs completed request check
If it is a _replayed_ request, we should always send a simple ack if it is
completed, because the client doesn't not care about any additional caps.

If it is a _resent_ request, then we want to return useful caps on open or
create requests, even if any modification side-effects have already been
committed.  The additional checks for completed already exist in the
create and open handlers.

Signed-off-by: Sage Weil <sage@newdream.net>